requirements.txt
discord.py
discord.py-pagination